refactor: Rename `lib/plugins/aws/utils/findAndGroupDeployments.js`
