+++ lib/plugins/aws/deploy-list.js
@@ -2,7 +2,7 @@
 
 const { log, writeText } = require('@serverless/utils/log');
 const validate = require('./lib/validate');
-const findAndGroupDeployments = require('./utils/findAndGroupDeployments');
+const findAndGroupDeployments = require('./utils/find-and-group-deployments');
 const setBucketName = require('./lib/set-bucket-name');
 
 class AwsDeployList {
