+++ lib/plugins/aws/deploy/lib/cleanupS3Bucket.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const _ = require('lodash');
-const findAndGroupDeployments = require('../../utils/findAndGroupDeployments');
+const findAndGroupDeployments = require('../../utils/find-and-group-deployments');
 const getS3ObjectsFromStacks = require('../../utils/getS3ObjectsFromStacks');
 const { log } = require('@serverless/utils/log');
