+++ lib/plugins/aws/rollback.js
@@ -6,7 +6,7 @@ const monitorStack = require('./lib/monitor-stack');
 const waitForChangeSetCreation = require('./lib/wait-for-change-set-creation');
 const getCreateChangeSetParams = require('./lib/get-create-change-set-params');
 const getExecuteChangeSetParams = require('./lib/get-execute-change-set-params');
-const findAndGroupDeployments = require('./utils/findAndGroupDeployments');
+const findAndGroupDeployments = require('./utils/find-and-group-deployments');
 const ServerlessError = require('../../serverless-error');
 const { style, log, progress } = require('@serverless/utils/log');
